Split Search city to 2 files
